initial swash to move BSplineKit, MixedModels and RobustModels to it's own weak dependency
