+++ Project.toml
@@ -34,6 +34,16 @@ Tables = "bd369af6-aec1-5ad0-b16a-f7cc5008161c"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 TimerOutputs = "a759f4b9-e2f1-59dc-863e-4aeb61b1ea8f"
 
+[weakdeps]
+BSplineKit = "093aae92-e908-43d7-9660-e50ee39d5a0a"
+MixedModels = "ff71e718-51f3-5ec2-a782-8ffcbfa3c316"
+RobustModels = "d6ea1423-9682-4bbd-952f-b1577cbf8c98"
+
+[extensions]
+UnfoldMixedModelsExt = "MixedModels"
+UnfoldRobustModelsExt = "RobustModels"
+UnfoldBSplineKitExt = "BSplineKit"
+
 [compat]
 BSplineKit = "0.16"
 CategoricalArrays = "0.10"
