+++ src/effects.jl
@@ -176,11 +176,7 @@ function cast_referenceGrid(r,eff,times;basisname=nothing)
 
     return result
 end
-Effects._trmequal(t1::Unfold.AbstractSplineTerm,t2::AbstractTerm) = _symequal(t1.term,t2)
-Effects._trmequal(t1::Unfold.AbstractSplineTerm,t2::Unfold.AbstractSplineTerm) = _symequal(t1.term,t2.term)
 
-Effects._trmequal(t1::AbstractTerm,t2::Unfold.AbstractSplineTerm) = _symequal(t1,t2.term)
-Effects._symequal(t1::AbstractTerm,t2::Unfold.AbstractSplineTerm) = _symequal(t1,t2.term)
 #Effects._symequal(t1::AbstractTerm,t2::Unfold.TimeExpandedTerm) = _symequal(t1,t2.term)
 #function Effects._replace(matrix_term::MatrixTerm{<:Tuple{<:Unfold.TimeExpandedTerm}},typicals::Dict)
