+++ src/timeexpandedterm.jl
@@ -19,14 +19,6 @@ struct TimeExpandedTerm{T<:AbstractTerm} <: AbstractTerm
 end
 
 
-function TimeExpandedTerm(
-    term::NTuple{N,Union{<:AbstractTerm,<:MixedModels.RandomEffectsTerm}},
-    basisfunction,
-    eventfields::Array{Symbol,1},
-) where {N}
-    # for mixed models, apply it to each Term
-    TimeExpandedTerm.(term, Ref(basisfunction), Ref(eventfields))
-end
 function TimeExpandedTerm(term, basisfunction, eventfields::Nothing)
     # if the eventfield is nothing, call default
     TimeExpandedTerm(term, basisfunction)
