+++ src/typedefinitions.jl
@@ -11,14 +11,6 @@ function DesignMatrix()
     return DesignMatrix([], [], [])
 end
 
-struct UnfoldMixedModelFitCollection{T<:AbstractFloat} <:
-       MixedModels.MixedModelFitCollection{T}
-    fits::Vector
-    λ::Vector{<:Union{LowerTriangular{T,Matrix{T}},Diagonal{T,Vector{T}}}}
-    inds::Vector{Vector{Int}}
-    lowerbd::Vector{T}
-    fcnames::NamedTuple
-end
 
 abstract type UnfoldModel end
 
@@ -64,7 +56,7 @@ UnfoldLinearMixedModelContinuousTime(d::Dict) =
 UnfoldLinearMixedModelContinuousTime(d::Dict, X::DesignMatrix) =
     UnfoldLinearMixedModelContinuousTime(d, X, [])
 
-abstract type AbstractSplineTerm <:AbstractTerm end
+
 abstract type ModelFit end
 
 struct LinearModelFit <: ModelFit
